# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Invoice Capture MVP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/vendors:
    get:
      tags: [analytics]
      operationId: getVendorAnalytics
      summary: Per-vendor invoice analytics
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendor analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorAnalyticsResponse"

  # ─── Phase 2: Exception Management ───────────────────────────────────────────
  /exceptions:
    get:
      tags: [exceptions]
      operationId: listExceptions
      summary: List invoices in exception with management filters
      parameters:
        - name: reason
          in: query
          schema:
            type: string
          description: Filter by exception reason substring
        - name: owner
          in: query
          schema:
            type: string
        - name: reviewed
          in: query
          schema:
            type: ["boolean", "null"]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [age, vendorName, totalAmount, confidenceScore, status]
            default: age
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Exception list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
  /invoices/{id}/exception/events:
    get:
      tags: [exceptions]
      operationId: getExceptionEvents
      summary: Exception activity history for an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exception events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExceptionEvent"
  /invoices/{id}/exception/note:
    post:
      tags: [exceptions]
      operationId: addExceptionNote
      summary: Add an internal note to an exception
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExceptionNoteInput"
      responses:
        "201":
          description: Note recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExceptionEvent"
  /invoices/{id}/exception/assign:
    post:
      tags: [exceptions]
      operationId: assignException
      summary: Assign an owner to an exception
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExceptionAssignInput"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/exception/review:
    post:
      tags: [exceptions]
      operationId: reviewException
      summary: Mark an exception as reviewed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExceptionReviewInput"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/exception/return-to-approval:
    post:
      tags: [exceptions]
      operationId: returnExceptionToApproval
      summary: Return a resolved exception to the approval queue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExceptionReturnInput"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "422":
          description: Invoice is not in a returnable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    Invoice:
      type: object
      required:
        - id
        - status
        - createdAt
        - updatedAt
        - fileObjectPath
        - originalFileName
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [PENDING_EXTRACTION, EXCEPTION, PENDING_APPROVAL, APPROVED, POSTED, VOIDED]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        invoiceNumber:
          type: ["string", "null"]
        invoiceDate:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        taxAmount:
          type: ["number", "null"]
        poNumber:
          type: ["string", "null"]
        currency:
          type: string
          default: USD
        fileObjectPath:
          type: string
        originalFileName:
          type: string
        documentId:
          type: ["string", "null"]
        businessDocumentId:
          type: ["string", "null"]
          description: Business-facing display ID formatted as "VendorID - InvoiceNumber - Amount" (e.g. "V-00123 - INV-12345 - 1309.00"). Null until vendor is matched and invoice number/total are available.
        vendorRawName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        voucherId:
          type: ["string", "null"]
        exceptionReason:
          type: ["string", "null"]
        lowConfidenceFields:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        subtotal:
          type: ["number", "null"]
        freightAmount:
          type: ["number", "null"]
        discountAmount:
          type: ["number", "null"]
          description: Discount/credit reduction extracted from the invoice (stored as a positive magnitude; subtracted in the tie-out formula).
        otherChargesAmount:
          type: ["number", "null"]
          description: Other charges, fees, or surcharges extracted from the invoice (added in the tie-out formula; negative values represent credits).
        tieOutExpectedTotal:
          type: ["number", "null"]
          description: Computed expected total = subtotal + tax + freight + other charges − discount.
        tieOutDifference:
          type: ["number", "null"]
          description: Extracted invoice total minus the expected total.
        tieOutStatus:
          type: ["string", "null"]
          description: Header tie-out outcome — one of PASS, WARNING, FAIL, SKIPPED.
        tieOutExplanation:
          type: ["string", "null"]
          description: Human-readable explanation of how the tie-out was computed and why it passed, warned, failed, or was skipped.
        paymentTerms:
          type: ["string", "null"]
        vendorMatchScore:
          type: ["number", "null"]
        validationStatus:
          type: ["string", "null"]
          description: Outcome of the authoritative validation engine (e.g. PASS, NEEDS_REVIEW, FAIL).
        reviewStatus:
          type: ["string", "null"]
          description: Review flag set when an invoice needs human review (e.g. NEEDS_REVIEW).
        overallReviewStatus:
          type: ["string", "null"]
          description: Aggregate human review state (e.g. PENDING, APPROVED).
        duplicateCheck:
          type: ["string", "null"]
          description: Result of the duplicate (vendor + invoice number) check.
        vendorCheck:
          type: ["string", "null"]
          description: Result of the vendor validation check (required/active/not on hold/matched).
        poCheck:
          type: ["string", "null"]
          description: Result of the purchase order capture check.
        amountCheck:
          type: ["string", "null"]
          description: Result of the amount validation check (total > 0, currency).
        totalTieOut:
          type: ["string", "null"]
          description: Result of the header tie-out check (subtotal + tax + freight = total).
        validationDetails:
          type: ["string", "null"]
          description: JSON detail of all validation checks, blocking issues, and warnings.
        extractionStatus:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED]
          default: PENDING
        extractionError:
          type: ["string", "null"]
        extractionAttempts:
          type: integer
          default: 0
        extractionErrorDetail:
          type: ["string", "null"]
          description: Safe JSON troubleshooting detail for a failed extraction (no sensitive data).
        fieldConfidence:
          type: ["string", "null"]
          description: JSON map of per-field confidence (0-100) keyed by field name.
        extractionNotes:
          type: ["string", "null"]
        lastExtractedAt:
          type: ["string", "null"]
          format: date-time
        sourceDocumentId:
          type: ["integer", "null"]
          description: ID of the source document this invoice was detected/split from.
        invoiceSequence:
          type: ["integer", "null"]
          description: 1-based position of this invoice within its source document.
        pageStart:
          type: ["integer", "null"]
          description: 1-based first page of this invoice within the source document.
        pageEnd:
          type: ["integer", "null"]
          description: 1-based last page (inclusive) of this invoice within the source document.
        role:
          type: string
          enum: [AP_PROCESSOR, AP_APPROVER]
          default: AP_PROCESSOR
        removedAt:
          type: ["string", "null"]
          format: date-time
          description: Timestamp when the invoice was voided/removed. Null when active.
        removedBy:
          type: ["string", "null"]
          description: Actor who voided/removed the invoice.
        removalReason:
          type: ["string", "null"]
          description: Required reason captured when the invoice was voided/removed.
        removalNote:
          type: ["string", "null"]
          description: Optional free-text note captured at removal time.
        exportStatus:
          type: ["string", "null"]
          description: File-based export readiness — one of NOT_READY, READY, EXPORTED, FAILED, BLOCKED. This is NOT an ERP status.
        exportBatchId:
          type: ["string", "null"]
          description: Business ID of the export batch this invoice was last included in.
        exportedAt:
          type: ["string", "null"]
          format: date-time
        exportBlockedReason:
          type: ["string", "null"]
        exportRetryCount:
          type: integer
          default: 0
        exportFileName:
          type: ["string", "null"]
        exportFormat:
          type: ["string", "null"]
        exceptionOwner:
          type: ["string", "null"]
          description: AP user assigned to work this exception.
        exceptionReviewedAt:
          type: ["string", "null"]
          format: date-time
        exceptionReviewedBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InvoiceListResponse:
      type: object
      required:
        - data
        - total
        - page
        - limit
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Phase 2: Dashboard & Analytics ──────────────────────────────────────
...
    VendorAnalyticsRow:
      type: object
      required:
        - vendorId
        - vendorCode
        - vendorName
        - invoiceCount
        - totalAmount
        - exceptionCount
        - exportedCount
      properties:
        vendorId:
          type: integer
        vendorCode:
          type: string
        vendorName:
          type: string
        vendorStatus:
          type: string
          description: ACTIVE, INACTIVE, or ON_HOLD.
        invoiceCount:
          type: integer
        totalAmount:
          type: number
        avgVendorMatchConfidence:
          type: ["number", "null"]
        exceptionCount:
          type: integer
        duplicateWarningCount:
          type: integer
          default: 0
        tieOutFailCount:
          type: integer
          default: 0
        missingPoCount:
          type: integer
          default: 0
        exportedCount:
          type: integer
    VendorAnalyticsResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/VendorAnalyticsRow"

    # ─── Phase 2: Exception Management ───────────────────────────────────────
    ExceptionEvent:
      type: object
      required: [id, invoiceId, eventType, createdAt]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        eventType:
          type: string
          description: NOTE, ASSIGNED, REVIEWED, RETURNED_TO_APPROVAL, or REOPENED.
        note:
          type: ["string", "null"]
        actor:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ExceptionNoteInput:
      type: object
      required: [note]
      properties:
        note:
          type: string
          minLength: 1
        actor:
          type: ["string", "null"]
    ExceptionAssignInput:
      type: object
      required: [owner]
      properties:
        owner:
          type: string
          minLength: 1
        actor:
          type: ["string", "null"]
    ExceptionReviewInput:
      type: object
      properties:
        reviewed:
          type: boolean
          default: true
        note:
          type: ["string", "null"]
        actor:
          type: ["string", "null"]
    ExceptionReturnInput:
      type: object
      properties:
        note:
          type: ["string", "null"]
        actor:
          type: ["string", "null"]

    # ─── Phase 2: Source Document list ───────────────────────────────────────
...