# lib/api-zod/src/generated/types/exceptionAssignInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Invoice Capture MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface ExceptionAssignInput {
  /** @minLength 1 */
  owner: string;
  /** @nullable */
  actor?: string | null;
}