# lib/api-zod/src/generated/types/exceptionReturnInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Invoice Capture MVP API
 * OpenAPI spec version: 0.1.0
 */

export interface ExceptionReturnInput {
  /** @nullable */
  note?: string | null;
  /** @nullable */
  actor?: string | null;
}