# lib/api-zod/src/generated/types/vendorAnalyticsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Invoice Capture MVP API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorAnalyticsRow } from './vendorAnalyticsRow';

export interface VendorAnalyticsResponse {
  data: VendorAnalyticsRow[];
}